Relations between user and projects

+++ app/Http/Controllers/ProjectController.php
@@ -4,15 +4,16 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Project;
+use Illuminate\Support\Facades\Auth;
 
 class ProjectController extends Controller
 {
     public function Index()
     {
-        //Retrieve all data the Projects table
-        $projects = Project::all();
+        //Retrieve all relative data for authenticated user's id
+        $projects = Project::where('user_id', Auth::id())->get();
 
-        //pass down data to view
+        //pass down data to view(in projects table)
         return view('account/projects/index', compact('projects'));
     }
 
@@ -25,9 +26,10 @@ class ProjectController extends Controller
     {
         $project = new Project();
 
-        $project::create(
-            $request->all()
-        );
+        $project::create([
+            'title' => $request->title,
+            'user_id' => Auth::id()
+        ]);
 
         return redirect("account/projects");
     }
@@ -51,9 +53,12 @@ class ProjectController extends Controller
 
     public function update(Request $request, $id)
     {
-        //Update project title 
-        Project::where('id', $id)
-            ->update(["title" => $request->title]);
+        //Update project title if project id matches
+        //& Right user is logged in(won't allow changes if it's wrong user)
+        Project::where('id', $id)->where(
+            'user_id',
+            Auth::id()
+        )->update(["title" => $request->title]);
         // return $id;
         return back();
     }
@@ -62,9 +67,13 @@ class ProjectController extends Controller
     {
         //Find one project where 'id' is equal to same id from route
         $project = Project::where('id', $id)->first();
-        //Delete Project and inspirations within it
-        $project->deleteRelated();
 
-        return redirect('account/projects');
+        if ($project->user_id == Auth::id()) {
+            //Delete Project and inspirations within it
+            $project->deleteRelated();
+            return redirect('account/projects');
+        } else {
+            return redirect('account/projects');
+        }
     }
 }
